Suppress warning if base treatment is default scalar
If multiple treatments are specified, a warning is given if a scalar is
provided as treatment value. Previously, since this check currently
included the base treatment T0, for which the default is 0, a warning
was emitted even if the base treatment was not specified.

Now, suppress misleading warnings for the base treatment being 0.