Suppress scalar warning if treatment is 0
If multiple treatments are specified, a warning is given if a scalar is
provided as treatment value. Previously, since this check currently
included the base treatment T0, for which the default is 0, a warning
was emitted even if the base treatment was not specified.

Now, to avoid misleading warnings for the base treatment, suppress the
warning if T is 0.