Improve discreteness handling, allow binary outcomes (#816)
Adds a binary_outcome keyword arg to most estimators, where if True then the outcome nuisance model will be a classifier.

Additionally add constraints to ensure nuisance model discreteness is handled appropriately by the user. 
If a nuisance model has a continuous target but a classifier is passed, then will raise an AttributeError.
Conversely, if a nuisance model has a discrete target but a regressor is passed, then a warning is issued.